+++ src/main/java/com/github/firewolf8385/customitemapi/items/attributes/attributes/ToughnessAttribute.java
@@ -17,7 +17,7 @@ public class ToughnessAttribute extends ItemAttribute {
     }
 
     @Override
-    public void addedToItem(CustomItem customItem, ItemBuilder item, int value) {
+    public void addedToItem(CustomItem customItem, ItemBuilder item, double value) {
         Attribute attribute = Attribute.GENERIC_ARMOR_TOUGHNESS;
 
         switch (customItem.getType()) {
